+++ AtomicCore.BlockChain.OmniscanAPI/JsonModel/Property/OmniPropertyCoinBasicJson.cs
@@ -65,7 +65,7 @@ namespace AtomicCore.BlockChain.OmniscanAPI
         /// propertyid
         /// </summary>
         [JsonProperty("propertyid")]
-        public int PropertyId { get; set; }
+        public ulong PropertyId { get; set; }
 
         /// <summary>
         /// subcategory
