+++ AtomicCore.BlockChain.OmniscanAPI/JsonModel/Transaction/OmniTransactionJson.cs
@@ -65,7 +65,7 @@ namespace AtomicCore.BlockChain.OmniscanAPI
         /// propertyid
         /// </summary>
         [JsonProperty("propertyid")]
-        public int Propertyid { get; set; }
+        public ulong Propertyid { get; set; }
 
         /// <summary>
         /// propertyname
